requirements.txt
Flask>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0
Jinja2>=3.1.4
